mcp: finalize MCP server definition provider API
Closes #243522
